+++ C/utils/history-stack.h
@@ -6,7 +6,7 @@
 
 typedef struct history_stack_struct history_stack;
 
-history_stack* history_stack_create(uint8_t size);
+history_stack* history_stack_create(uint16_t size);
 
 void history_stack_push(history_stack* stack, void* item);
